requirements.txt
streamlit[auth]>=1.37.0
streamlit-oauth>=0.1.14
toml>=0.10.2
requests>=2.31.0
cryptography>=41.0.0
python-dotenv>=1.0.0
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0